More thoroughly unfuck varargs.
The method as it was would only work if there had been one level of fucked varargs. But if we're
going to encourage a world in which you can pass an array in for varags and expect it to work, then
we should be able to expect it work even if it happens through a series of 2+ varags method calls.